refactor: final wallet id stuff
